Split visitCode
